refactoring: move old_addon check into separate file
